Search ancestor directories for the bakefile
